Moved CV components under components directory
