+++ sto_sister/components/icon_slot_detector.py
