+++ sto_sister/components/label_locator.py
